==> 1st_Publication/03_JAGS_Hierarchical_Models.Rmd <==
---
title: "The Independent model, Bayesian hierarchical models for Country-specific, Region, Income and Income-Region"
author: "Oluwayomi Akinfenwa"
date: "08/03/2024"
output:
  html_document:
    toc: yes
    toc_depth: '4'
    df_print: paged
  bookdown::html_document2:
    toc: no
    toc_float: yes
    toc_depth: 4
    number_sections: yes
  pdf_document:
    toc: yes
    toc_depth: '4'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(knitr.duplicate.label = 'allow')
```

Sourcing the file containing all the data needed for this analysis from the Rmd file named: 01_PISA_Analysis_Data.Rmd.
This file contains the raw data from OECD website and all the necessary data for this analysis.

```{r loading the data}
sys.source("01_JAGS_PISA_Analysis_Data.R", envir = knitr::knit_global())
```


### Setting the model specification for the JAGS Models

```{r Ind_Countrymodel}
if (!dir.exists(here::here("Manuscript_Models")))
  dir.create(here::here("Manuscript_Models"))

if(file.exists(here::here("Manuscript_Models",
                          "CountryInd_JAGSModel.Rdata"))){
  load(here::here("Manuscript_Models",
                          "CountryInd_JAGSModel.Rdata"))
  countryJags <- FALSE
} else countryJags <- TRUE
```

```{r, eval = countryJags}
#For the Independent linear model, we are interested in fitting a linear model omitting the countries with one data point.
#Hence, we will use the SPISA_Europe_Data for this model fitting.

Ind_C_JAGS_Model = "
model{
# Data model
for (i in 1:n){
y.i[i] ~ dnorm(mu.i[i], sigma^-2)

mu.i[i] <- alpha.j[C[i]] + beta.j[C[i]]*(x.i[i])
}

# Prior on country alphas and betas

for(j in 1:n_C){
alpha.j[j] ~ dnorm(500, 100^-2)
beta.j[j] ~ dnorm(0, 5^-2)
}


# Prior on sigma
sigma ~ dt (30, 10^-2,1)T(0,) # Truncated t-distribution

# Setting up the posterior
for(m in 1:n_pred){
ytilde[m] ~ dnorm(mu_pred.m[m], sigma^-2)
mu_pred.m[m] <- alpha.j[C_pred[m]] + beta.j[C_pred[m]]*(xpred.m[m])
}

# Setting up yrep for posterior predictive checks
for (i in 1:n){
yrep[i] ~ dnorm(mu.i[i], sigma^-2)
}
}
"

## JAGS Model
Ind_C_Jags.data <- list(y.i = SPisa_Europe_Data$math,
                    x.i = SPisa_Europe_Data$year_orig,
                    n = nrow(SPisa_Europe_Data),
                    n_C = SPisa_Europe_Data$Country |> unique() |> length(),
                    C = as.numeric(factor(SPisa_Europe_Data$Country)),
                    n_pred = nrow(Spred_grid),
                    C_pred = Spred_grid$Country_num,
                    xpred.m = Spred_grid$year_orig)

Ind_C_Parnames <- c("alpha.j","beta.j","sigma","ytilde", "yrep","mu_pred.m","mu.i")

# model
CountryInd_JAGSModel <- jags(data = Ind_C_Jags.data,
              parameters.to.save = Ind_C_Parnames,
              model.file = textConnection(Ind_C_JAGS_Model),
              n.iter = 10000,
              n.burnin = 2000,
              n.chains = 4,
              n.thin = 4)

save(CountryInd_JAGSModel, file=here::here("Manuscript_Models", "CountryInd_JAGSModel.Rdata"))
```

#### Country-specific model

```{r Countrymodel}
if(file.exists(here::here("Manuscript_Models",
                          "Country_JAGSModel.Rdata"))){
  load(here::here("Manuscript_Models",
                          "Country_JAGSModel.Rdata"))
  countryJags <- FALSE
} else countryJags <- TRUE
```

```{r, eval = countryJags}
C_JAGS_Model = "
model{
# Data model
for (i in 1:n){
y.i[i] ~ dnorm(mu.i[i], sigma^-2)

mu.i[i] <- alpha.j[C[i]] + beta.j[C[i]]*(x.i[i])
}

# Prior on country alphas

for(j in 1:n_C){
alpha.j[j] ~ dnorm(mu_alpha, sigma_alpha^-2)
beta.j[j] ~ dnorm(mu_beta, sigma_beta^-2)
}

# Prior on sigma, model variation
sigma ~ dt (30, 10^-2,1)T(0,) # Truncated t-distribution
mu_alpha ~ dnorm(500, 100^-2)
mu_beta ~ dnorm(0, 5^-2)
sigma_alpha ~ dt(30, 10^-2,1)T(0,) # Truncated t-distribution
sigma_beta ~ dt( 0, 2^-2,1)T(0,) # Truncated t-distribution 


# Setting up the posterior
for(m in 1:n_pred){
ytilde[m] ~ dnorm(mu_pred.m[m], sigma^-2)
mu_pred.m[m] <- alpha.j[C_pred[m]] + beta.j[C_pred[m]] *(xpred.m[m])
}

# Setting up yrep for posterior predictive  checks
for (i in 1:n){
yrep[i] ~ dnorm(mu.i[i], sigma^-2)
}
}
"

## JAGS Model
C_Jags.data <- list(y.i = PISA_Europe_Data$math,
                    x.i = PISA_Europe_Data$year_orig,
                    n = nrow(PISA_Europe_Data),
                    n_C = PISA_Europe_Data$Country |> unique() |> length(),
                    C = as.numeric(factor(PISA_Europe_Data$Country)),
                    n_pred = nrow(pred_grid),
                    C_pred = pred_grid$Country_num,
                    xpred.m = pred_grid$year_orig)

C_Parnames <- c("alpha.j","beta.j", "mu_alpha", "mu_beta","sigma","sigma_alpha",
                 "sigma_beta","ytilde", "yrep","mu_pred.m","mu.i")

# model
Country_JAGSModel <- jags(data = C_Jags.data,
              parameters.to.save = C_Parnames,
              model.file = textConnection(C_JAGS_Model),
              n.iter = 10000,
              n.burnin = 2000,
              n.chains = 4,
              n.thin = 4)

save(Country_JAGSModel, file=here::here("Manuscript_Models", "Country_JAGSModel.Rdata"))
```

**This C_MODEL fits a separate linear regression for each country, giving each country its distinct parameter estimates.**

# Hierarchical Model

## Region-based Country specific intercept and slope

### Region as a hierarchical structure.

```{r Regionmodel}
if(file.exists(here::here("Manuscript_Models",
                          "Region_JAGSModel.Rdata"))){
  load(here::here("Manuscript_Models",
                          "Region_JAGSModel.Rdata"))
  regionJags <- FALSE
} else regionJags <- TRUE
```

```{r, eval = regionJags}
hmregion = "
model{
# Data model
for (i in 1:n){
y.i[i] ~ dnorm(mu.i[i], sigma^-2)

mu.i[i] <- alpha.j[C[i]] + beta.j[C[i]]*(x.i[i] )
}

# Prior on alphas to vary by region

for(j in 1:n_C){
alpha.j[j] ~ dnorm(mu_alpha.r[R[j]], sigma_alpha[R[j]]^-2)
beta.j[j] ~ dnorm(mu_beta.r[R[j]], sigma_beta^-2)
}

for (r in 1:n_R){
mu_alpha.r[r] ~ dnorm(mu_alpha_eur, sigma_alpha.r^-2)
mu_beta.r[r] ~ dnorm(mu_beta_eur, sigma_beta.r^-2)
sigma_alpha[r] ~ dt( 30, 10^-2,1)T(0,) # Truncated t-distribution
}

# Prior on sigma, model variation
sigma ~ dt (30, 10^-2,1)T(0,) # Truncated t-distribution

# Prior on alpha hyper-parameters
mu_alpha_eur ~ dnorm(500, 100^-2)
sigma_alpha.r ~ dt( 30,10^-2,1)T(0,) # Truncated t-distribution

# Prior on beta hyper-parameters
mu_beta_eur ~ dnorm(0, 5^-2)
sigma_beta ~ dt( 0, 2^-2,1)T(0,) # Truncated t-distribution 
sigma_beta.r ~ dt( 0, 2^-2,1)T(0,) # Truncated t-distribution


for(m in 1:n_pred){
ytilde[m] ~ dnorm(mu_pred.m[m], sigma^-2)
mu_pred.m[m] <- alpha.j[C_pred[m]] + beta.j[C_pred[m]] *(xpred.m[m])
}

# Setting up yrep for posterior predictive  checks
for (i in 1:n){
yrep[i] ~ dnorm(mu.i[i], sigma^-2)
}
}
"

## JAGS Model

R_Jags.data <- list(y.i = PISA_Europe_Data$math,
                  x.i = PISA_Europe_Data$year_orig,
                  n = nrow(PISA_Europe_Data),
                  n_C = PISA_Europe_Data$Country %>% unique() %>% length(),
                  C = as.numeric(factor(PISA_Europe_Data$Country)),
                  R = country_region$region_num,
                  n_pred = nrow(pred_grid),
                  C_pred = pred_grid$Country_num,
                  xpred.m = pred_grid$year_orig,
                  n_R = PISA_Europe_Data$Region %>% unique() %>% length())

R_Parnames <- c( "alpha.j","beta.j","sigma","sigma_alpha","sigma_beta","mu_alpha.r", 
               "mu_beta.r","mu_alpha_eur","mu_beta_eur", "ytilde", "yrep",
               "mu_pred.m", "mu.i" ,"sigma_alpha.r", "sigma_beta.r")

# model
Region_JAGSModel <- jags(data = R_Jags.data,
              parameters.to.save = R_Parnames,
              model.file = textConnection(hmregion),
              n.iter = 10000,
              n.burnin = 2000,
              n.chains = 4,
              n.thin = 4)

save(Region_JAGSModel, file=here::here("Manuscript_Models", "Region_JAGSModel.Rdata"))
```

**R_Model fits a hierarchical model on the PISA_Europe_Data by estimating the model parameters based on the estimates from its different level of hierarchy (geographical region)**

### Income as a hierarchical structure.

```{r Incomemodel}
if(file.exists(here::here("Manuscript_Models",
                          "Income_JAGSModel.Rdata"))){
  load(here::here("Manuscript_Models",
                          "Income_JAGSModel.Rdata"))
  incomeJags <- FALSE
} else incomeJags <- TRUE
```

```{r, eval = incomeJags}
hmincome = "
model{
# Data model
for (i in 1:n){
y.i[i] ~ dnorm(mu.i[i], sigma^-2)

mu.i[i] <- alpha.j[C[i]] + beta.j[C[i]]*(x.i[i])
}

# Prior on alphas to vary by income

for(j in 1:n_C){
alpha.j[j] ~ dnorm(mu_alpha.I[I[j]], sigma_alpha[I[j]]^-2)
beta.j[j] ~ dnorm(mu_beta.I[I[j]], sigma_beta^-2)
}

for (r in 1:n_I){
mu_alpha.I[r] ~ dnorm(mu_alpha_inc, sigma_alpha.inc^-2)
mu_beta.I[r] ~ dnorm(mu_beta_inc, sigma_beta.inc^-2)
sigma_alpha[r] ~ dt( 30, 10^-2,1)T(0,) # Truncated t-distribution
}

# Prior on sigma, model variation
sigma ~ dt (30, 10^-2,1)T(0,) # Truncated t-distribution

# Prior on alpha hyper-parameters
mu_alpha_inc ~ dnorm(500, 100^-2)
sigma_alpha.inc ~ dt( 30, 10^-2,1)T(0,) # Truncated t-distribution

# Prior on beta hyper-parameters
mu_beta_inc ~ dnorm(0, 5^-2)
sigma_beta ~ dt( 0, 2^-2,1)T(0,) # Truncated t-distribution
sigma_beta.inc ~ dt( 0, 2^-2,1)T(0,) # Truncated t-distribution


for(m in 1:n_pred){
ytilde[m] ~ dnorm(mu_pred.m[m], sigma^-2)
mu_pred.m[m] <- alpha.j[C_pred[m]] + beta.j[C_pred[m]] *(xpred.m[m] - mean(xpred.m))
}

# Setting up yrep for posterior predictive  checks
for (i in 1:n){
yrep[i] ~ dnorm(mu.i[i], sigma^-2)
}
}
"

## JAGS Model

I_Jags.data <- list(y.i = PISA_Europe_Data$math,
                    x.i = PISA_Europe_Data$year_orig,
                    n = nrow(PISA_Europe_Data),
                    n_C = PISA_Europe_Data$Country %>% unique() %>% length(),
                    C = as.numeric(factor(PISA_Europe_Data$Country)),
                    I = country_income$income_num,
                    n_pred = nrow(pred_grid),
                    C_pred = pred_grid$Country_num,
                    xpred.m = pred_grid$year_orig,
                    n_I = PISA_Europe_Data$Income %>% unique() %>% length())

I_Parnames <- c( "alpha.j","beta.j","sigma","sigma_alpha","sigma_beta","mu_alpha.I", 
                 "mu_beta.I","mu_alpha_inc","mu_beta_inc", "ytilde", "yrep",     "mu_pred.m", "mu.i" ,"sigma_alpha.inc", "sigma_beta.inc")
 
# model
Income_JAGSModel <- jags(data = I_Jags.data,
                parameters.to.save = I_Parnames,
                model.file = textConnection(hmincome),
                n.iter = 10000,
                n.burnin = 2000,
                n.chains = 4,
                n.thin = 4)

save(Income_JAGSModel, file=here::here("Manuscript_Models", "Income_JAGSModel.Rdata"))
```

**I_Model fits a hierarchical model on the PISA_Europe_Data by estimating the model parameters based on the estimates from its different level of hierarchy (income classification)**

### Income_Region as a hierarchical structure.

```{r IncomeRegionmodel}
if(file.exists(here::here("Manuscript_Models",
                          "IncomeRegion_JAGSModel.Rdata"))){
  load(here::here("Manuscript_Models",
                          "IncomeRegion_JAGSModel.Rdata"))
  incomeregionJags <- FALSE
} else incomeregionJags <- TRUE
```

```{r, eval=incomeregionJags}
hmincome_region = "
model{
# Data model
for (i in 1:n){
y.i[i] ~ dnorm(mu.i[i], sigma^-2)

mu.i[i] <- alpha.j[C[i]] + beta.j[C[i]]*(x.i[i] - mean(x.i))
}

# Prior on alphas to vary by income

for(j in 1:n_C){
alpha.j[j] ~ dnorm(mu_alpha.IR[IR[j]], sigma_alpha[IR[j]]^-2)
beta.j[j] ~ dnorm(mu_beta.IR[IR[j]], sigma_beta^-2)
}

for (r in 1:n_IR){
mu_alpha.IR[r] ~ dnorm(mu_alpha_inre, sigma_alpha.inre^-2)
mu_beta.IR[r] ~ dnorm(mu_beta_inre, sigma_beta.inre^-2)
sigma_alpha[r] ~ dt( 0, 100^-2,1)T(0,) # Truncated t-distribution
}

# Prior on sigma, model variation
sigma ~ dt (30, 10^-2,1)T(0,) # Truncated t-distribution

# Prior on alpha hyper-parameters
mu_alpha_inre ~ dnorm(500, 100^-2)
sigma_alpha.inre ~ dt( 0, 100^-2,1)T(0,) # Truncated t-distribution

# Prior on beta hyper-parameters
mu_beta_inre ~ dnorm(0, 5^-2)
sigma_beta ~ dt( 0, 2^-2,1)T(0,) # Truncated t-distribution
sigma_beta.inre ~ dt( 0, 2^-2,1)T(0,) # Truncated t-distribution


for(m in 1:n_pred){
ytilde[m] ~ dnorm(mu_pred.m[m], sigma^-2)
mu_pred.m[m] <- alpha.j[C_pred[m]] + beta.j[C_pred[m]] *(xpred.m[m] - mean(xpred.m))
}

# Setting up yrep for posterior predictive  checks
for (i in 1:n){
yrep[i] ~ dnorm(mu.i[i], sigma^-2)
}
}
"
IR_Jags.data <- list(y.i = PISA_Europe_Data$math,
                    x.i = PISA_Europe_Data$year_orig,
                    n = nrow(PISA_Europe_Data),
                    n_C = PISA_Europe_Data$Country %>% unique() %>% length(),
                    C = as.numeric(factor(PISA_Europe_Data$Country)),
                    IR = country_incomeregion$incomeregion_num,
                    n_pred = nrow(pred_grid),
                    C_pred = pred_grid$Country_num,
                    xpred.m = pred_grid$year_orig,
                    n_IR = PISA_Europe_Data$Income_Region %>% unique() %>% length())

IR_Parnames <- c("alpha.j","beta.j","sigma","sigma_alpha","sigma_beta","mu_alpha.IR", 
                 "mu_beta.IR","mu_alpha_inre","mu_beta_inre", "ytilde", "yrep", "mu_pred.m", 
                 "mu.i" ,"sigma_alpha.inre", "sigma_beta.inre")
# model
IR_JAGSModel <- jags(data = IR_Jags.data,
                parameters.to.save = IR_Parnames,
                model.file = textConnection(hmincome_region),
                n.iter = 10000,
                n.burnin = 2000,
                n.chains = 4,
                n.thin = 4)

save(IR_JAGSModel, file=here::here("Manuscript_Models", "IncomeRegion_JAGSModel.Rdata"))
```
